% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateBalance_function.R
\name{CalculateBalance}
\alias{CalculateBalance}
\title{Calculate balance of covariates before and after matching.}
\usage{
CalculateBalance(dtaBef, dtaAfter = NULL, cols, trt,
  diff_means = FALSE)
}
\arguments{
\item{dtaBef}{Data frame including the data before matching.}

\item{dtaAfter}{Data frame including only the matched pairs data. Can be set to NULL if
we want to calculate the standardized difference of means for only one
dataset.}

\item{cols}{Column indeces of the variables we want to check. Should be the same in
both dtaBef and dtaAfter}

\item{trt}{Column index for treatment variable.}

\item{diff_means}{Logical. Set equal to TRUE to return the difference in
means of treated versus controls. Defaults to FALSE.}
}
\value{
A 2 by length(cols) matrix with the standardized difference of
means before and after matching. (Second row is empty if dtaAfter is
not given.)
}
\description{
Function that calculates the standardized difference of means.
}
