% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPS_functions.R
\name{DAPSest}
\alias{DAPSest}
\title{Fitting DAPSm with fixed or optimal weight}
\usage{
DAPSest(dataset, out.col = NULL, trt.col = NULL, caliper = 0.1,
  weight = "optimal", coords.columns = NULL, ignore.cols = NULL,
  pairsRet = FALSE, cov.cols = NULL, cutoff = 0.1, w_tol = 0.01,
  distance = StandDist, caliper_type = c("DAPS", "PS"), quiet = FALSE,
  coord_dist = FALSE, true_value = NULL)
}
\arguments{
\item{dataset}{Data frame including treatment, outcome, coordinates, propensity score
estimates (named prop.scores) and observed confounders.}

\item{out.col}{If outcome column name is not 'Y', out.col should be the index of the outcome
column.}

\item{trt.col}{If treatment is not named 'X', trt.col should be set to the index of the
treatment column.}

\item{caliper}{A caliper for the DAPS Score difference of matched pairs. Defaults to 0.1.}

\item{weight}{Scalar between 0 and 1 or should be set to 'optimal'. Describes the percent
of weight to be given on PS difference. 1 - weight is given to standardized
distance. If set to 'optimal', the 'optimal' weight is chosen. Defaults to
'optimal'.}

\item{coords.columns}{If the columns of coordinates are not named 'Longitude' and 'Latitude',
coords.columns are the column indices corresponding to longitude and latitude
accordingly.}

\item{pairsRet}{Whether we want to return the information on the matched pairs. Logical.
Defaults to FALSE.}

\item{cov.cols}{If the weight is set to 'optimal', standardized difference of means will be
calculated on the columns whose indices are in cov.cols.}

\item{cutoff}{The cutoff of standardized difference of means under which the covariates are
considered balanced. Defaults to 0.1.}

\item{w_tol}{Tolerance on the choice of the optimal weight. Only needed when weight is
'optimal'. Defaults to 0.01.}

\item{distance}{Function that takes in the distance matrix and returns the standardized
distance matrix. Defaults to the funcion that subtracks the minimum and
divides by the range.}

\item{caliper_type}{Whether we want the caliper to be on DAPS or on the PS. caliper_type must be
either 'DAPS', or 'PS'.}

\item{quiet}{Whether we want to print the choice of weight in DAPS optimal. Defauls to TRUE.}

\item{coord_dist}{Set to true when we want to use a distance function that calculates the
spherical distance of points instead of Euclidean. Defaults to FALSE.}

\item{true_value}{Numeric. If provided, an indicator of whether the CI covers the true value is
returned.}

\item{ignore.col}{All column indices that should not be included in the linear model. Often,
this should be set to all columns indices corresponding to columns other than
outcome, treatment, and observed confounders.}
}
\value{
A list including: the estimate of the causal effect, and potential
standardized difference of means, optimal weight chosen, information on matched
pairs.
}
\description{
Function that fits DAPS by choosing the 'optimal' weight, or using the weight
specified and returns the causal effect estimate under a specification of
caliper. If the weight is set to 'optimal', the optimal weight is chosen as
the smallest weight that satisfies standardized difference less than a cutoff.
For computational reasons, we do not fit DAPS on a lot of weights, but perform
an iterative algorithm to identify the optimal w.
}
\examples{

}

