% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weight_balance_functions.R
\name{DAPSWeightCE}
\alias{DAPSWeightCE}
\title{Plot the effect estimate as a function of w.}
\usage{
DAPSWeightCE(dataset, out.col, weights, pairs, chosen_w)
}
\arguments{
\item{dataset}{The dataset that was supplied to CalcDAPSWeightBalance() for calculating
balance.}

\item{out.col}{The index of the outcome column if it is not named 'Y' in the dataset.}

\item{weights}{The weights that we used to fit DAPSm.}

\item{pairs}{A list where each element corresponds to a weight. Each element is a vector
including the row indices of the dataset that are included in the matched
dataset for each weight w. 2nd element of the list returned by
CalcDAPSWeightBalance().}

\item{chosen_w}{The weight value that was chosen by DAPSchoiceModel().}
}
\description{
Plotting the effect estimate from various fit of DAPSm for varying w. The
chosen w will be the only red dot. A loess curve is fit to the effect
estimates.
}

