% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_functions.R
\name{EmpCDF}
\alias{EmpCDF}
\title{Standardizing a matrix to be between 0 and 1 using the empirical CDF.}
\usage{
EmpCDF(x)
}
\arguments{
\item{x}{Matrix to be standardized.}
}
\value{
A matrix of the same dimensions as the one given.
}
\description{
Takes in a matrix and returns a matrix of the same dimension with entries
equal to the corresponding empirical cdf value.
}
\examples{
set.seed(1)
D <- matrix(rexp(800, rate = 0.5), 20, 40)
Dnew <- EmpCDF(D)
hist(D)
hist(Dnew)
}
