% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormDataset.R
\name{FormDataset}
\alias{FormDataset}
\title{Forming a data set.}
\usage{
FormDataset(dataset, ignore.cols = NULL, out.col = NULL, trt.col = NULL)
}
\arguments{
\item{dataset}{The data frame we want to reform.}

\item{ignore.cols}{The indices of the columns we want to drop. Defaults
to NULL. If set to NULL, no columns are dropped.}

\item{out.col}{The index of the outcome column which will be renamed to
'Y'. Defaults to NULL. If set to NULL, no column will be renamed.}

\item{trt.col}{The index of the treatment column which will be renamed
to 'X'. Defaults to NULL. If set to NULL, no column will be renamed.}
}
\value{
A data frame with renamed the columns of treatment and outcome,
and no longer includes some of the columns.
}
\description{
Takes in a data frame and changes the names of the treatment and outcome
columns, and drops unnecessary ones.
}
\examples{
D <- cbind(Trt = rbinom(100, 1, 1/2), Out = rnorm(100),
           C1 = rnorm(100), C2 = rnorm(100))
D_new <- FormDataset(D, ignore.cols = 4, out.col = 2, trt.col = 1)
}
