% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchedDataMap.R
\name{MatchedDataMap}
\alias{MatchedDataMap}
\title{Plot matches.}
\usage{
MatchedDataMap(x, trt_coords, con_coords, plot.title = "",
  point_data = TRUE)
}
\arguments{
\item{x}{A data frame where the number of rows corresponds to the number
of matches. There are at least four columns in the data frame and correspond
to longitude and latitude information of the treated and the control unit.
Within the same row, we must have the coordinate information of treated and
control that are matched to each other.}

\item{trt_coords}{Indeces of the columns including the longitude and latitude
(with this order) of the treated unit.}

\item{con_coords}{Indeces of the columns including the longitude and latitude
(with this order) of the control unit.}

\item{point_data}{Whether we want to print points of the treated and control
observations. Treated units are green points and control units are red points.}

\item{plot_title}{Title of the plot.}
}
\value{
Plot of matched pairs.
}
\description{
Plots a map of the US of the matched pairs with color corresponding to
which unit is treated and which is the control and lines connecting
matched pairs.
}
\examples{
data('toyData2')
toyData2$prop.scores <- glm(Z ~ X1 + X2 + X3 + X4, family = binomial,
                            data = toyData2)$fitted.values
daps1 <- DAPSest(toyData2, out.col = 2, trt.col = 1, caliper = 0.5,
                 weight = 0.2, coords.columns = c(4, 5),
                 pairsRet = TRUE, cov.cols = 6:9, cutoff = 0.1,
                 w_tol = 0.001, coord_dist = TRUE, caliper_type = 'DAPS')
MatchedDataMap(x = daps1$pairs, trt_coords = c(3, 4), con_coords = c(7, 8))
# For a larger weight, less weight is given to distance, so matches will
# further apart.
daps2 <- DAPSest(toyData2, out.col = 2, trt.col = 1, caliper = 0.5,
                 weight = 0.8, coords.columns = c(4, 5),
                 pairsRet = TRUE, cov.cols = 6:9, cutoff = 0.1,
                 w_tol = 0.001, coord_dist = TRUE, caliper_type = 'DAPS')
MatchedDataMap(x = daps2$pairs, trt_coords = c(3, 4), con_coords = c(7, 8))
}
