% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPS_functions.R
\name{MinDistMatch}
\alias{MinDistMatch}
\title{Matching on distance matrix}
\usage{
MinDistMatch(M, caliper = NULL)
}
\arguments{
\item{M}{A non-negative matrix (T x C)}

\item{caliper}{The caliper controls how far apart with respect to their M entry
matched pairs can be. When set to NULL, all possible matches with
finite M entry are taken. If set to a number, no entries are
allowed to be matched if larger than caliper.}
}
\value{
An ordered matrix of the row and column indices that are matched.
}
\description{
Takes in a matrix and runs through it matching rows to columns with small entries.
The smallest one is matched first, and it continues on.
}
\examples{
set.seed(1)
D <- matrix(rexp(800, rate = 0.5), 20, 40)
MinDistMatch(D, caliper = NULL)
MinDistMatch(D, caliper = 0.1)

}

