% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_plot_functions.R
\name{PlotBalance}
\alias{PlotBalance}
\title{Plotting covariate balance before and after matching.}
\usage{
PlotBalance(dtaBef, dtaAfter, cols, trt, cutoff = 0.1)
}
\arguments{
\item{dtaBef}{Data frame include the data before matching.}

\item{dtaAfter}{Data frame including only the matched pairs data. Can be set to
NULL if we want to plot the standardized difference of means for only one dataset.}

\item{cols}{Column indeces of the variables we want to check.}

\item{trt}{Column index for treatment variable.}

\item{cutoff}{Cutoff for standardized difference. Defaults to 0.1.}
}
\value{
Plot of standardized difference of means before and after matching.
}
\description{
Function that plots the standardized difference of means before and after matching.
}

