% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_functions.R
\name{StandDist}
\alias{StandDist}
\title{Standardizing a matrix to be between 0 and 1.}
\usage{
StandDist(x)
}
\arguments{
\item{x}{Matrix to be standardized. There should be at least two different
entries in x.}
}
\value{
A matrix of the same dimensions as the one given.
}
\description{
Takes in a matrix and subtract the minimum element and divides by the range
to return a matrix with elements between 0 and 1.
}
\examples{
set.seed(1)
D <- matrix(rexp(800, rate = 0.5), 20, 40)
Dnew <- StandDist(D)
hist(D)
hist(Dnew)
}

