% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPS_functions.R
\name{MinDistMatch}
\alias{MinDistMatch}
\title{Linear regression}
\usage{
MinDistMatch(M, caliper = NULL)
}
\arguments{
\item{y}{response vector (1 x n)}

\item{X}{covariate matrix (p x n) with no intercept}
}
\value{
A list with 4 elements: coefficients, vcov, sigma, df
}
\description{
Runs an OLS regression not unlike \code{\link{lm}}
}
\examples{
data(mtcars)
X <- as.matrix(mtcars[, c("cyl", "disp", "hp")])
y <- mtcars[, "mpg"]
linreg(y, X)

}

